refactored vulkan instance, device, and swapchain

+++ include/G10/GXtypedef.h
@@ -18,12 +18,14 @@ typedef          char      i8;
 typedef unsigned short     u16;
 typedef signed   short     s16;
 typedef          short     i16;
-typedef unsigned long      u32;
-typedef signed   long      s32;
-typedef          long      i32;
+typedef unsigned int       u32;
+typedef signed   int       s32;
+typedef          int       i32;
 typedef unsigned long long u64;
 typedef signed   long long s64;
 typedef          long long i64;
+
+
 
 // Floats
 typedef float              f32;
